requirements.txt
keras~=2.7.0rc2
pandas~=1.3.4
numpy~=1.19.5
progress~=1.6
sklearn~=0.0
scikit-learn~=1.0
matplotlib~=3.4.3
typing~=3.7.4.3
Pillow~=8.4.0
opencv-python~=4.5.4.58
tensorflow~=2.7.0rc1